requirements.txt
absl-py==2.1.0
activate==0.2.1
aiofiles==23.2.1
aiosignal==1.3.1
appnope==0.1.4
astunparse==1.6.3
asynctest==0.13.0
attrs==24.2.0
autogen==1.0.16
autopep8==2.0.4
backcall==0.2.0
backoff==2.2.1
base==1.0.4
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.0.0
blis==0.7.11
cached-property==1.5.2
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==3.1.0
chroma-hnswlib==0.7.3
click==8.1.7
coloredlogs==15.0.1
cycler==0.10.0
cymem==2.0.8
debugpy==1.7.0
decorator==5.1.1
distlib==0.3.6
distro==1.9.0
dnspython==2.3.0
docopt==0.6.2
docutils==0.20.1
entrypoints==0.4
exceptiongroup==1.2.2
fastapi==0.103.2
ffmpy==0.3.2
filelock==3.9.0
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gast==0.4.0
google-api-core==2.19.0
google-auth==2.27.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-speech==2.26.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
gradio==3.34.0
gradio_client==0.2.6
graphlib_backport==1.1.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
h5py==3.8.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==4.12.0
keras==2.11.0
keyring==24.1.1
kiwisolver==1.2.0
langcodes==3.3.0
libclang==16.0.6
linkify-it-py==2.0.3
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
nest-asyncio==1.6.0
numphy==0.0.2
numpy==1.21.6
oauthlib==3.2.2
onnxruntime==1.14.1
openai==1.39.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
orjson==3.9.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.1
pandas==1.3.5
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pkginfo==1.10.0
pkgutil_resolve_name==1.3.10
platformdirs==3.1.0
posthog==3.5.0
preshed==3.0.9
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.24.4
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pydantic==1.10.15
pydub==0.25.1
Pygments==2.17.2
pymongo==4.3.3
pyparsing==2.4.7
PyPika==0.48.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-multipart==0.0.8
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.2
readme-renderer==37.3
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rsa==4.9
schedule==1.2.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
selenium==4.11.2
semantic-version==2.10.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.4.1
source==1.2.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.27.0
sympy==1.10.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
thinc==8.2.3
threadpoolctl==3.1.0
tokenizers==0.20.0
tomli==2.0.1
toolz==0.12.1
torch==1.13.1
torchaudio==0.13.1
torchtext==0.4.0
torchvision==0.14.1
tornado==6.2
tqdm==4.66.2
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.11.1
twine==4.0.2
typer==0.9.4
uc-micro-py==1.0.3
urllib3==1.26.15
utils==1.0.2
uvicorn==0.22.0
uvloop==0.18.0
virtualenv==20.20.0
wasabi==1.1.2
watchfiles==0.20.0
wcwidth==0.2.13
weasel==0.3.4
webdriver-manager==4.0.2
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.15.0
